resume_portfolio/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
black==23.3.0
boto3==1.26.105
botocore==1.29.105
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Django==4.2
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-compressor==4.3.1
django-cors-headers==4.1.0
django-csp==3.7
django-environ==0.10.0
django-filter==23.1
django-guardian==2.4.0
django-sass-processor==1.2.2
django-user-agents==0.4.0
djangorestframework==3.14.0
docopt==0.6.2
folium==0.14.0
fonttools==4.39.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Js2Py==0.74
kiwisolver==1.4.4
libsass==0.22.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.6.0.66
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
Pillow==9.5.0
pipreqs==0.4.13
pipwin==0.5.2
platformdirs==3.2.0
plotly==5.14.1
pyalchemy==0.0.7
Pygments==2.14.0
pyjsparser==2.7.1
pylibmc==1.6.3
pymemcache==4.0.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
pysqlite3==0.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rcssmin==1.1.1
requests==2.28.2
rjsmin==1.2.1
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.8.1
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlparse==0.4.3
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
yarg==0.1.9
zipp==3.15.0
